Make merge and remove more robust

+++ Templates/CondorSubmit/removeRootFiles.sh
@@ -1,2 +1,4 @@
 #!/bin/bash
-rm $(find . -regex '.*_.*_[0-9]+\.root')
+if [[ -f $1 ]]; then
+    rm $(find . -regex '.*_.*_[0-9]+\.root')
+fi
